HW-0 redo create project & add git_ignore
