--- jjunji/README.md
@@ -1,1 +1,0 @@
-### 전지훈 공간입니다.
