+++ jjunji/StudyApplication/app/.gitignore
