+++ jjunji/StudyApplication/app/src/main/java/com/example/studyapplication/MainActivity.java
@@ -1,0 +1,14 @@
+package com.example.studyapplication;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.os.Bundle;
+
+public class MainActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_main);
+    }
+}
